refactor: ♻️ extract meeting list components and time utilities
