+++ src/routes/summary/+page.svelte
@@ -1,10 +1,9 @@
 <script lang="ts">
-  import { LightSwitch, RadioItem, RadioGroup, TabGroup, Tab } from "@skeletonlabs/skeleton";
+  import { LightSwitch, TabGroup, Tab } from "@skeletonlabs/skeleton";
 
   import GroupList from "$lib/components/summary/GroupsCarousel.svelte";
-  import CalendarIcon from "~icons/material-symbols/calendar-month";
-  import LocationIcon from "~icons/material-symbols/location-on";
-  import ClockIcon from "~icons/material-symbols/nest-clock-farsight-analog-outline";
+  import ScheduledMeetingsList from "$lib/components/summary/ScheduledMeetingsList.svelte";
+  import UnscheduledMeetingsList from "$lib/components/summary/UnscheduledMeetingsList.svelte";
 
   let tabSet: number = 0;
 
@@ -154,59 +153,6 @@
     endTime: string;
     location: string;
   };
-
-  // Sort meetings by startDate
-  scheduledMeetings.sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime());
-  unscheduledMeetings.sort(
-    (a, b) => new Date(a.startDate).getTime() - new Date(b.startDate).getTime(),
-  );
-
-  // Group meetings by startDate
-  let groupedMeetingsByDate: Record<string, ScheduledMeeting[]> = {};
-  scheduledMeetings.forEach((meeting) => {
-    const date = meeting.date;
-    if (!groupedMeetingsByDate[date]) {
-      groupedMeetingsByDate[date] = [];
-    }
-    groupedMeetingsByDate[date].push(meeting);
-  });
-
-  function convertIsoToDate(isoDateString: string): string {
-    const options: Intl.DateTimeFormatOptions = {
-      month: "short",
-      day: "numeric",
-    };
-
-    const date = new Date(isoDateString);
-    return date.toLocaleDateString("en-US", { ...options, timeZone: "UTC" });
-  }
-
-  function convertIsoToWeekdayDate(isoDateString: string): string {
-    const options: Intl.DateTimeFormatOptions = {
-      weekday: "short",
-      month: "short",
-      day: "numeric",
-    };
-
-    const date = new Date(isoDateString);
-    return date.toLocaleDateString("en-US", { ...options, timeZone: "UTC" });
-  }
-
-  function convertTo12HourFormat(time: string): string {
-    const [hours, minutes] = time.split(":");
-    let period = "am";
-
-    let hours12 = parseInt(hours, 10);
-
-    if (hours12 >= 12) {
-      period = "pm";
-      if (hours12 > 12) {
-        hours12 -= 12;
-      }
-    }
-
-    return `${hours12}:${minutes} ${period}`;
-  }
 </script>
 
 <LightSwitch />
@@ -229,103 +175,10 @@
 
       <svelte:fragment slot="panel">
         {#if tabSet === 0}
-          <div class="flex flex-col gap-4">
-            {#each Object.keys(groupedMeetingsByDate) as date}
-              <div class="p-2 card variant-glass">
-                <h2 class="mb-2 text-xl font-bold md:text-2xl">{convertIsoToWeekdayDate(date)}</h2>
-
-                <div class="flex flex-col gap-2">
-                  {#each groupedMeetingsByDate[date] as meeting}
-                    <div
-                      class="flex flex-col justify-between gap-4 p-3 bg-center bg-cover rounded-lg md:items-center h-fit md:flex-row card hover:variant-ghost"
-                    >
-                      <div class="flex flex-wrap items-center justify-between w-full gap-2">
-                        <div class="flex flex-col gap-y-1">
-                          <a href={meeting.link} target="_blank" referrerpolicy="no-referrer">
-                            <p class="text-xl font-bold md:text-2xl line-clamp-1 max-h-12">
-                              {meeting.name}
-                            </p>
-                          </a>
-
-                          <div class="flex flex-row flex-wrap gap-x-4">
-                            <p class="flex items-center gap-1 text-md md:text-lg">
-                              <ClockIcon />
-                              {convertTo12HourFormat(meeting.startTime)} - {convertTo12HourFormat(
-                                meeting.endTime,
-                              )}
-                            </p>
-                            <p class="flex items-center gap-1 text-md md:text-lg">
-                              <LocationIcon />
-                              {meeting.location}
-                            </p>
-                          </div>
-                        </div>
-                      </div>
-
-                      <RadioGroup
-                        class="flex items-center h-fit w-fit"
-                        active="variant-filled-primary"
-                        hover="hover:variant-soft-primary"
-                      >
-                        <RadioItem
-                          bind:group={meeting.attending}
-                          name="justify"
-                          value={"Yes"}
-                          required>Yes</RadioItem
-                        >
-                        <RadioItem
-                          bind:group={meeting.attending}
-                          name="justify"
-                          value={"No"}
-                          required>No</RadioItem
-                        >
-                        <RadioItem
-                          bind:group={meeting.attending}
-                          name="justify"
-                          value={"Maybe"}
-                          required>Maybe</RadioItem
-                        >
-                      </RadioGroup>
-                    </div>
-                  {/each}
-                </div>
-              </div>
-            {/each}
-          </div>
+          <div class="flex flex-col gap-4"><ScheduledMeetingsList {scheduledMeetings} /></div>
         {:else if tabSet === 1}
           <div class="flex flex-col gap-2">
-            {#each unscheduledMeetings as meeting}
-              <div
-                class="flex flex-col justify-between gap-4 p-3 bg-center bg-cover rounded-lg md:items-center h-fit md:flex-row card hover:variant-ghost"
-              >
-                <div class="flex flex-wrap items-center justify-between w-full gap-2">
-                  <div class="flex flex-col gap-y-1">
-                    <a href={meeting.link} target="_blank" referrerpolicy="no-referrer">
-                      <p class="text-xl font-bold md:text-2xl line-clamp-1 max-h-12">
-                        {meeting.name}
-                      </p>
-                    </a>
-
-                    <div class="flex flex-row flex-wrap gap-x-4">
-                      <p class="flex items-center gap-1 text-md md:text-lg">
-                        <CalendarIcon />
-                        {convertIsoToDate(meeting.startDate)} - {convertIsoToDate(meeting.endDate)}
-                      </p>
-                      <p class="flex items-center gap-1 text-md md:text-lg">
-                        <ClockIcon />
-                        {convertTo12HourFormat(meeting.startTime)} - {convertTo12HourFormat(
-                          meeting.endTime,
-                        )}
-                      </p>
-                      <p class="flex items-center gap-1 text-md md:text-lg">
-                        <LocationIcon />
-                        {meeting.location}
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            {/each}
+            <UnscheduledMeetingsList {unscheduledMeetings} />
           </div>
         {/if}
       </svelte:fragment>
